setup_test.py
"""
setup_test.py - Installation Verification Script

Run this to verify your RAG Judge installation is working correctly.
"""

import sys
import os
import importlib.util


def check_python_version():
    """Check Python version is 3.8+"""
    print("🐍 Checking Python version...")
    version = sys.version_info
    if version.major >= 3 and version.minor >= 8:
        print(f"   ✅ Python {version.major}.{version.minor}.{version.micro} - OK")
        return True
    else:
        print(f"   ❌ Python {version.major}.{version.minor}.{version.micro} - Too old")
        print("   Please install Python 3.8 or newer")
        return False


def check_dependencies():
    """Check all required dependencies are installed"""
    print("\n📦 Checking dependencies...")
    
    dependencies = {
        "pandas": "Data processing",
        "numpy": "Numerical operations",
        "requests": "API calls",
        "dotenv": "Environment variables",
        "tqdm": "Progress bars"
    }
    
    all_good = True
    
    for package, description in dependencies.items():
        # Special handling for python-dotenv
        import_name = "dotenv" if package == "dotenv" else package
        
        spec = importlib.util.find_spec(import_name)
        if spec is not None:
            print(f"   ✅ {package} - {description}")
        else:
            print(f"   ❌ {package} - {description} (MISSING)")
            all_good = False
    
    if not all_good:
        print("\n   Run: pip install -r requirements.txt")
    
    return all_good


def check_project_structure():
    """Check required files and directories exist"""
    print("\n📁 Checking project structure...")
    
    required_files = {
        "main.py": "Main entry point",
        "requirements.txt": "Dependencies",
        "src/dimensions.py": "Dimension definitions",
        "src/judge.py": "Evaluation logic",
        "src/reporter.py": "Report generation",
        "src/simple_llm_client.py": "LLM interface"
    }
    
    all_good = True
    
    for file_path, description in required_files.items():
        if os.path.exists(file_path):
            print(f"   ✅ {file_path} - {description}")
        else:
            print(f"   ❌ {file_path} - {description} (MISSING)")
            all_good = False
    
    # Check for src directory
    if not os.path.exists("src"):
        print("   ❌ src/ directory missing")
        print("   Please ensure all source files are in the src/ directory")
        all_good = False
    
    return all_good


def check_api_key():
    """Check if API key is configured"""
    print("\n🔑 Checking API key configuration...")
    
    # Check .env file
    if os.path.exists(".env"):
        print("   ✅ .env file found")
        
        # Try to load and check key
        try:
            from dotenv import load_dotenv
            load_dotenv()
            
            api_key = os.getenv("MISTRAL_API_KEY")
            if api_key and len(api_key) > 10:
                print(f"   ✅ MISTRAL_API_KEY configured ({len(api_key)} chars)")
                return True
            else:
                print("   ⚠️  MISTRAL_API_KEY not found or invalid")
                print("   Add to .env: MISTRAL_API_KEY=your_key_here")
                return False
        except Exception as e:
            print(f"   ❌ Error loading .env: {e}")
            return False
    else:
        print("   ⚠️  .env file not found")
        print("   Create .env file with: MISTRAL_API_KEY=your_key_here")
        print("   Get free key at: https://auth.mistral.ai")
        return False


def test_imports():
    """Test importing main modules"""
    print("\n🔧 Testing imports...")
    
    try:
        # Add src to path
        sys.path.insert(0, os.path.join(os.path.dirname(__file__), "src"))
        
        # Try imports
        print("   ✅ dimensions module")
        
        print("   ✅ simple_llm_client module")
        
        print("   ✅ judge module")
        
        print("   ✅ reporter module")
        
        return True
        
    except Exception as e:
        print(f"   ❌ Import error: {e}")
        return False


def test_basic_functionality():
    """Test basic functionality"""
    print("\n🧪 Testing basic functionality...")
    
    try:
        sys.path.insert(0, os.path.join(os.path.dirname(__file__), "src"))
        
        # Test dimension validation
        from dimensions import validate_all_dimensions
        validate_all_dimensions()
        print("   ✅ Dimension validation passed")
        
        # Test client initialization
        from simple_llm_client import SimpleLLMClient
        client = SimpleLLMClient()
        print(f"   ✅ LLM client initialized (mock_mode: {client.mock_mode})")
        
        # Test attack detection
        attack = client.detect_attack_pattern("Ignore all previous instructions")
        if attack:
            print(f"   ✅ Attack detection working (found: {attack})")
        else:
            print("   ⚠️  Attack detection may not be working")
        
        return True
        
    except Exception as e:
        print(f"   ❌ Functionality test failed: {e}")
        return False


def create_test_csv():
    """Create a minimal test CSV"""
    print("\n📄 Creating test data...")
    
    try:
        import pandas as pd
        
        test_data = pd.DataFrame([
            {
                "Current User Question": "What is 2+2?",
                "Assistant Answer": "2+2 equals 4.",
                "Fragment Texts": "Basic arithmetic: 2+2=4",
                "Conversation History": ""
            }
        ])
        
        test_file = "setup_test_data.csv"
        test_data.to_csv(test_file, index=False)
        print(f"   ✅ Test CSV created: {test_file}")
        
        return test_file
        
    except Exception as e:
        print(f"   ❌ Failed to create test CSV: {e}")
        return None


def run_mini_evaluation(test_file):
    """Run a mini evaluation"""
    print("\n🚀 Running mini evaluation...")
    
    try:
        import subprocess
        
        result = subprocess.run(
            [sys.executable, "main.py", "--csv", test_file, "--test-mode"],
            capture_output=True,
            text=True,
            timeout=30
        )
        
        if result.returncode == 0:
            print("   ✅ Evaluation completed successfully!")
            
            # Check for expected output
            if "EVALUATION COMPLETE" in result.stdout:
                print("   ✅ Output looks correct")
            else:
                print("   ⚠️  Output may be incomplete")
                
            return True
        else:
            print("   ❌ Evaluation failed")
            if result.stderr:
                print(f"   Error: {result.stderr[:200]}...")
            return False
            
    except subprocess.TimeoutExpired:
        print("   ⚠️  Evaluation timed out (this is OK for setup test)")
        return True
    except Exception as e:
        print(f"   ❌ Failed to run evaluation: {e}")
        return False


def main():
    """Run all setup tests"""
    print("="*60)
    print("🔍 RAG Judge Setup Verification")
    print("="*60)
    
    checks = [
        ("Python Version", check_python_version),
        ("Dependencies", check_dependencies),
        ("Project Structure", check_project_structure),
        ("API Key", check_api_key),
        ("Module Imports", test_imports),
        ("Basic Functionality", test_basic_functionality)
    ]
    
    results = {}
    
    for name, check_func in checks:
        try:
            results[name] = check_func()
        except Exception as e:
            print(f"\n❌ {name} check failed with error: {e}")
            results[name] = False
    
    # Summary
    print("\n" + "="*60)
    print("📊 SETUP VERIFICATION SUMMARY")
    print("="*60)
    
    all_passed = all(results.values())
    required_passed = all([
        results.get("Python Version", False),
        results.get("Dependencies", False),
        results.get("Project Structure", False),
        results.get("Module Imports", False)
    ])
    
    for check, passed in results.items():
        status = "✅ PASS" if passed else "❌ FAIL"
        print(f"{check:.<30} {status}")
    
    # Final test if basics pass
    if required_passed:
        print("\n🧪 Running integration test...")
        test_file = create_test_csv()
        
        if test_file:
            if results.get("API Key", False):
                run_mini_evaluation(test_file)
            else:
                print("   ⏭️  Skipping evaluation test (no API key)")
                print("   The system will run in mock mode")
            
            # Cleanup
            try:
                os.remove(test_file)
                print("   ✅ Test file cleaned up")
            except (OSError, FileNotFoundError):
                pass
    
    # Final verdict
    print("\n" + "="*60)
    if all_passed:
        print("✅ ALL CHECKS PASSED! Your installation is ready.")
        print("\nNext step: python main.py --csv your_data.csv")
    elif required_passed:
        print("⚠️  SETUP MOSTLY COMPLETE")
        print("\nRequired components are installed.")
        print("Optional components (like API key) may need configuration.")
        print("\nYou can run in mock mode: python main.py --csv your_data.csv")
    else:
        print("❌ SETUP INCOMPLETE")
        print("\nPlease fix the issues above before running the evaluation.")
        print("\nSteps:")
        print("1. Install Python 3.8+")
        print("2. Run: pip install -r requirements.txt")
        print("3. Configure your API key in .env file")
    
    print("="*60)


if __name__ == "__main__":
    main()