requirements.txt
# Core dependencies
pandas>=1.3.0,<2.0.0
numpy>=1.21.0,<1.25.0
python-dotenv>=0.19.0
requests>=2.26.0
tqdm>=4.62.0

# Optional dependencies for visualizations
# Uncomment if you want to generate charts
# matplotlib>=3.4.0
# seaborn>=0.11.0

# Development dependencies
# pytest>=6.2.0
# black>=21.9b0
# flake8>=3.9.0

# Type checking (optional)
# mypy>=0.910
# types-requests